vfs: rename namespace package to vfs
